+++ src/main/java/searchengine/controllers/ApiController.java
@@ -6,14 +6,16 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import searchengine.dto.indexing.IndexingResponse;
 import searchengine.repositories.PageRepository;
-import searchengine.services.indexing.*;
+import searchengine.repositories.SiteRepository;
+import searchengine.services.IndexingService;
 
 import searchengine.dto.search.SearchResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import searchengine.services.search.SearchService;
+import searchengine.services.RepositoryService;
+import searchengine.services.SearchService;
 import searchengine.dto.statistics.StatisticsResponse;
-import searchengine.services.statistics.StatisticsService;
+import searchengine.services.StatisticsService;
 
 @Slf4j
 @Setter
@@ -26,6 +28,8 @@ public class ApiController {
 	private final IndexingService indexingService;
 	private final StatisticsService statisticsService;
 	private final PageRepository pageRepository;
+	private final RepositoryService repositoryService;
+	private final SiteRepository siteRepository;
 
 	@GetMapping("/statistics")
 	public ResponseEntity<StatisticsResponse> statistics() {
@@ -59,7 +63,7 @@ public class ApiController {
 
 	@PostMapping("/test")
 	public ResponseEntity<?> test(@RequestParam final int id){
-		pageRepository.deleteById(id);
+		repositoryService.deleteSite(siteRepository.findById(id));
 		return new ResponseEntity<>(HttpStatus.ACCEPTED);
 	}
 }