+++ src/main/java/searchengine/services/Impl/StatisticsServiceImpl.java
@@ -1,4 +1,4 @@
-package searchengine.services.statistics;
+package searchengine.services.Impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -13,6 +13,7 @@ import searchengine.model.SiteEntity;
 import searchengine.repositories.LemmaRepository;
 import searchengine.repositories.PageRepository;
 import searchengine.repositories.SiteRepository;
+import searchengine.services.StatisticsService;
 import searchengine.tools.indexing.IndexingActions;
 
 import java.util.ArrayList;