+++ src/main/java/searchengine/services/IndexingService.java
@@ -1,13 +1,16 @@
-package searchengine.services.indexing;
+package searchengine.services;
+
 import org.springframework.http.ResponseEntity;
 import searchengine.dto.indexing.IndexingResponse;
 
 public interface IndexingService {
 
 	ResponseEntity<IndexingResponse> indexingStop();
+
 	ResponseEntity<IndexingResponse> indexingStart();
+
 	ResponseEntity<IndexingResponse> indexingPageStart(String url);
 
-	void setPressedStop(boolean value);
+//	void setPressedStop(boolean value);
 
 }