+++ src/main/java/searchengine/tools/indexing/IndexingActions.java
@@ -1,13 +1,15 @@
 package searchengine.tools.indexing;
-
 import searchengine.model.SiteEntity;
 import java.util.Set;
 
 public interface IndexingActions {
 
 	void startFullIndexing(Set<SiteEntity> siteEntities);
+
 	void startPartialIndexing(SiteEntity siteEntity);
+
 	void setIndexingActionsStarted(boolean value);
+
 	boolean getIndexingActionsStarted();
 
 }