+++ CHANGELOG.md
@@ -10,7 +10,8 @@ ___
 ### Added
 
 - [Batch](https://github.com/lebedev-artem/searchengine-master/blob/0623f920f6cdd3d09077b04e681452f4653e8c03/src/main/resources/application.yaml#L33) INSERT / UPDATE hibernate statement.
-
+- Cascade deletion all dependent tables by **site**.
+- In [application.yaml](https://github.com/lebedev-artem/searchengine-master/blob/f11a009daa1e161bb9ebc1fe85d767eef4ddc6ad/src/main/resources/application.yaml) added [properties](https://github.com/lebedev-artem/searchengine-master/blob/f11a009daa1e161bb9ebc1fe85d767eef4ddc6ad/src/main/resources/application.yaml#L76), that can be used to control certain actions with the table
 
 ### Fixed